+++ src/utils/Poller.ts
@@ -68,6 +68,7 @@ export default class Poller extends EventEmitter {
   }
 
   reset() {
+    console.info('Reset process poller');
     this.isWowRunning = false;
 
     if (this.child) {
@@ -77,6 +78,7 @@ export default class Poller extends EventEmitter {
   }
 
   start() {
+    console.info('Start process poller');
     this.reset();
     this.poll();
   }